Уменьшение связанности модулей
