+++ src/components/HomeView.ts
@@ -1,17 +1,13 @@
-import { UiConfig } from '../app/uiConfig';
-import { Product, ProductId } from '../types';
 import { Component } from '../ui/Component';
 import { ensureElement } from '../utils/utils';
-import { ProductView } from './ProductView';
 
 interface IHomeViewModel {
 	counter: number;
-	gallery: Product[];
+	gallery: HTMLElement[];
 	locked: boolean;
 }
 
 interface IHomeViewEvents {
-	onProductCardClick: (id: ProductId) => void;
 	onBasketOpenClick: () => void;
 }
 
@@ -37,20 +33,8 @@ export class HomeView extends Component<IHomeViewModel> {
 		this.setText(this._headerCounter, String(value));
 	}
 
-	set products(products: Product[]) {
-		const cards = products.map((product) => {
-			const productView = new ProductView(
-				UiConfig.templates.cardCatalogTemplate,
-				{
-					onProductCardClick: () => {
-						this.events.onProductCardClick(product.id);
-					},
-				},
-				'catalog'
-			);
-			return productView.render(product);
-		});
-		this._gallery.replaceChildren(...cards);
+	set gallery(elements: HTMLElement[]) {
+		this._gallery.replaceChildren(...elements);
 	}
 
 	set locked(value: boolean) {
